+++ plugin-gradle/src/main/java/com/diffplug/gradle/spotless/JavaExtension.java
@@ -21,18 +21,16 @@ import java.io.File;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
+import java.util.Map;
 import java.util.Objects;
 
 import javax.inject.Inject;
 
-import org.gradle.api.GradleException;
 import org.gradle.api.Project;
-import org.gradle.api.file.FileCollection;
-import org.gradle.api.plugins.JavaPluginConvention;
 import org.gradle.api.tasks.SourceSet;
 
 import com.diffplug.spotless.FormatterStep;
-import com.diffplug.spotless.extra.EclipseBasedStepBuilder;
+import com.diffplug.spotless.extra.EquoBasedStepBuilder;
 import com.diffplug.spotless.extra.java.EclipseJdtFormatterStep;
 import com.diffplug.spotless.generic.LicenseHeaderStep;
 import com.diffplug.spotless.java.CleanthatJavaStep;
@@ -42,7 +40,7 @@ import com.diffplug.spotless.java.ImportOrderStep;
 import com.diffplug.spotless.java.PalantirJavaFormatStep;
 import com.diffplug.spotless.java.RemoveUnusedImportsStep;
 
-public class JavaExtension extends FormatExtension implements HasBuiltinDelimiterForLicense {
+public class JavaExtension extends FormatExtension implements HasBuiltinDelimiterForLicense, JvmLang {
 	static final String NAME = "java";
 
 	@Inject
@@ -50,9 +48,7 @@ public class JavaExtension extends FormatExtension implements HasBuiltinDelimite
 		super(spotless);
 	}
 
-	// If this constant changes, don't forget to change the similarly-named one in
-	// testlib/src/test/java/com/diffplug/spotless/generic/LicenseHeaderStepTest.java as well
-	static final String LICENSE_HEADER_DELIMITER = "package ";
+	static final String LICENSE_HEADER_DELIMITER = LicenseHeaderStep.DEFAULT_JAVA_HEADER_DELIMITER;
 
 	@Override
 	public LicenseHeaderConfig licenseHeader(String licenseHeader) {
@@ -113,7 +109,11 @@ public class JavaExtension extends FormatExtension implements HasBuiltinDelimite
 
 	/** Removes any unused imports. */
 	public void removeUnusedImports() {
-		addStep(RemoveUnusedImportsStep.create(provisioner()));
+		addStep(RemoveUnusedImportsStep.create(RemoveUnusedImportsStep.defaultFormatter(), provisioner()));
+	}
+
+	public void removeUnusedImports(String formatter) {
+		addStep(RemoveUnusedImportsStep.create(formatter, provisioner()));
 	}
 
 	/** Uses the <a href="https://github.com/google/google-java-format">google-java-format</a> jar to format source code. */
@@ -199,14 +199,22 @@ public class JavaExtension extends FormatExtension implements HasBuiltinDelimite
 
 	public class PalantirJavaFormatConfig {
 		final String version;
+		String style;
 
 		PalantirJavaFormatConfig(String version) {
 			this.version = Objects.requireNonNull(version);
+			this.style = PalantirJavaFormatStep.defaultStyle();
 			addStep(createStep());
 		}
 
+		public PalantirJavaFormatConfig style(String style) {
+			this.style = Objects.requireNonNull(style);
+			replaceStep(createStep());
+			return this;
+		}
+
 		private FormatterStep createStep() {
-			return PalantirJavaFormatStep.create(version, provisioner());
+			return PalantirJavaFormatStep.create(version, style, provisioner());
 		}
 	}
 
@@ -219,7 +227,7 @@ public class JavaExtension extends FormatExtension implements HasBuiltinDelimite
 	}
 
 	public class EclipseConfig {
-		private final EclipseBasedStepBuilder builder;
+		private final EquoBasedStepBuilder builder;
 
 		EclipseConfig(String version) {
 			builder = EclipseJdtFormatterStep.createBuilder(provisioner());
@@ -232,6 +240,12 @@ public class JavaExtension extends FormatExtension implements HasBuiltinDelimite
 			Project project = getProject();
 			builder.setPreferences(project.files(configFiles).getFiles());
 			replaceStep(builder.build());
+		}
+
+		public EclipseConfig withP2Mirrors(Map<String, String> mirrors) {
+			builder.setP2Mirrors(mirrors);
+			replaceStep(builder.build());
+			return this;
 		}
 
 	}
@@ -361,15 +375,10 @@ public class JavaExtension extends FormatExtension implements HasBuiltinDelimite
 	@Override
 	protected void setupTask(SpotlessTask task) {
 		if (target == null) {
-			JavaPluginConvention javaPlugin = getProject().getConvention().findPlugin(JavaPluginConvention.class);
-			if (javaPlugin == null) {
-				throw new GradleException("You must either specify 'target' manually or apply the 'java' plugin.");
-			}
-			FileCollection union = getProject().files();
-			for (SourceSet sourceSet : javaPlugin.getSourceSets()) {
-				union = union.plus(sourceSet.getAllJava());
-			}
-			target = union;
+			target = getSources(getProject(),
+					"You must either specify 'target' manually or apply the 'java' plugin.",
+					SourceSet::getAllJava,
+					file -> file.getName().endsWith(".java"));
 		}
 
 		steps.replaceAll(step -> {