+++ plugin-maven/src/test/java/com/diffplug/spotless/maven/incremental/UpToDateCheckingTest.java
@@ -31,8 +31,10 @@ import com.diffplug.spotless.maven.MavenRunner;
 
 class UpToDateCheckingTest extends MavenIntegrationHarness {
 
+	private static final String DISABLED_MESSAGE = "Up-to-date checking disabled";
+
 	@Test
-	void upToDateCheckingDisabledByDefault() throws Exception {
+	void upToDateCheckingEnabledByDefault() throws Exception {
 		writePom(
 				"<java>",
 				"  <googleJavaFormat/>",
@@ -41,18 +43,29 @@ class UpToDateCheckingTest extends MavenIntegrationHarness {
 		List<File> files = writeUnformattedFiles(1);
 		String output = runSpotlessApply();
 
-		assertThat(output).doesNotContain("Up-to-date checking enabled");
+		assertThat(output).doesNotContain(DISABLED_MESSAGE);
 		assertFormatted(files);
 	}
 
 	@Test
-	void enableUpToDateChecking() throws Exception {
+	void explicitlyEnableUpToDateChecking() throws Exception {
 		writePomWithUpToDateCheckingEnabled(true);
 
 		List<File> files = writeUnformattedFiles(1);
 		String output = runSpotlessApply();
 
-		assertThat(output).contains("Up-to-date checking enabled");
+		assertThat(output).doesNotContain(DISABLED_MESSAGE);
+		assertFormatted(files);
+	}
+
+	@Test
+	void explicitlyDisableUpToDateChecking() throws Exception {
+		writePomWithUpToDateCheckingEnabled(false);
+
+		List<File> files = writeUnformattedFiles(1);
+		String output = runSpotlessApply();
+
+		assertThat(output).contains(DISABLED_MESSAGE);
 		assertFormatted(files);
 	}
 
@@ -63,7 +76,7 @@ class UpToDateCheckingTest extends MavenIntegrationHarness {
 		List<File> files = writeUnformattedFiles(1);
 		String output = runSpotlessApply();
 
-		assertThat(output).contains("Up-to-date checking enabled");
+		assertThat(output).doesNotContain(DISABLED_MESSAGE);
 		assertFormatted(files);
 	}
 
@@ -76,40 +89,7 @@ class UpToDateCheckingTest extends MavenIntegrationHarness {
 		List<File> files = writeUnformattedFiles(1);
 		String output = runSpotlessApply();
 
-		assertThat(output).contains("Up-to-date checking enabled");
-		assertFormatted(files);
-	}
-
-	private void writePomWithPluginManagementAndDependency() throws IOException {
-		setFile("pom.xml").toContent(createPomXmlContent("/pom-test-management.xml.mustache",
-				null,
-				null,
-				new String[]{
-						"<java>",
-						"  <googleJavaFormat/>",
-						"</java>",
-						"<upToDateChecking>",
-						"  <enabled>true</enabled>",
-						"</upToDateChecking>"},
-				new String[]{
-						"<dependencies>",
-						"  <dependency>",
-						"    <groupId>javax.inject</groupId>",
-						"    <artifactId>javax.inject</artifactId>",
-						"    <version>1</version>",
-						"  </dependency>",
-						"</dependencies>"},
-				null));
-	}
-
-	@Test
-	void disableUpToDateChecking() throws Exception {
-		writePomWithUpToDateCheckingEnabled(false);
-
-		List<File> files = writeUnformattedFiles(1);
-		String output = runSpotlessApply();
-
-		assertThat(output).doesNotContain("Up-to-date checking enabled");
+		assertThat(output).doesNotContain(DISABLED_MESSAGE);
 		assertFormatted(files);
 	}
 
@@ -124,7 +104,7 @@ class UpToDateCheckingTest extends MavenIntegrationHarness {
 		List<File> files = writeUnformattedFiles(1);
 		String output = runSpotlessApply();
 
-		assertThat(output).contains("Up-to-date checking enabled");
+		assertThat(output).doesNotContain(DISABLED_MESSAGE);
 		assertFormatted(files);
 		assertThat(indexFile.getParent()).exists();
 		assertThat(indexFile).exists();
@@ -143,7 +123,7 @@ class UpToDateCheckingTest extends MavenIntegrationHarness {
 		List<File> files = writeUnformattedFiles(1);
 		String output = runSpotlessApply();
 
-		assertThat(output).doesNotContain("Up-to-date checking enabled");
+		assertThat(output).contains(DISABLED_MESSAGE);
 		assertFormatted(files);
 		assertThat(indexFile.getParent()).exists();
 		assertThat(indexFile).doesNotExist();
@@ -213,6 +193,25 @@ class UpToDateCheckingTest extends MavenIntegrationHarness {
 
 		String checkOutput3 = runSpotlessCheck();
 		assertSpotlessCheckSkipped(files, checkOutput3);
+	}
+
+	private void writePomWithPluginManagementAndDependency() throws IOException {
+		setFile("pom.xml").toContent(createPomXmlContent("/pom-test-management.xml.mustache",
+				null,
+				null,
+				new String[]{
+						"<java>",
+						"  <googleJavaFormat/>",
+						"</java>"},
+				new String[]{
+						"<dependencies>",
+						"  <dependency>",
+						"    <groupId>javax.inject</groupId>",
+						"    <artifactId>javax.inject</artifactId>",
+						"    <version>1</version>",
+						"  </dependency>",
+						"</dependencies>"},
+				null));
 	}
 
 	private void writePomWithUpToDateCheckingEnabled(boolean enabled) throws IOException {